requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
attrs==25.3.0
binaryornot==0.4.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cookiecutter==2.6.0
decorator==5.2.1
fastapi==0.116.1
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
flet-web==0.28.3
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==2.2.1
multidict==6.6.4
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pillow==10.4.0
proglog==0.1.12
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.5
rich==14.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
ttkbootstrap==1.14.2
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==4.0.2
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
